requirements.txt
bcrypt==4.2.0
requests==2.32.5
python-jose[cryptography]==3.5.0
langgraph==1.0.3
langchain-openai==1.0.2
langchain-core==1.0.7
langchain==1.0.8
langgraph-supervisor==0.0.30
httpx==0.28.1
langfuse==3.8.1
pydantic==2.12.3
pillow==11.3.0
fastapi==0.119.0
streamlit-cookies-controller==0.0.4
typing-extensions==4.15.0
pandas==2.3.3
openpyxl==3.1.5
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
chromadb==1.3.4
python-dotenv==1.1.1
openai==2.6.1
passlib==1.7.4
markdown